add support for admin user role

+++ poll_app/app/auth.py
@@ -1,10 +1,15 @@
-from app import db
-from flask_login import login_user, logout_user
+from functools import wraps
+from app import db, login_manager
+from flask_login import login_user, logout_user, login_required, current_user
 from werkzeug.security import generate_password_hash, check_password_hash
 from app.models import User
 
 def load_user(user_id):
     return User.query.get(int(user_id))
+
+def get_all_users():
+    users = User.query.all()
+    return users
 
 # returns True if login attempt succeeds, else False
 def login(email, password):
@@ -34,8 +39,42 @@ def signup(email, password):
     db.session.commit()
     return True
 
-def delete_user(user_id):
-    logout_user()
+def delete_user(user_id, initiator='self'):
+    if (initiator == 'self'):
+        logout_user()
+    # otherwise it means admin deleted him
     user = User.query.get(user_id)
-    db.session.delete(user)
-    db.session.commit()
+    if user is not None and not user.is_admin:
+        db.session.delete(user)
+        db.session.commit()
+
+# if admin deletes a user while the user is still logged in, that user will be able to see polls until his session expires
+# however won't be able to create new polls or vote in polls
+
+
+
+
+
+def add_admins(admins):
+    for (admin_email, admin_pass) in admins.items():
+        
+        existing_user = User.query.filter_by(email=admin_email).first()
+        if existing_user:
+            print(f"User with email {admin_email} already exists. Skipping...")
+            continue  # Skip adding this admin
+
+        hashed_password = generate_password_hash(admin_pass)
+        new_user = User(email=admin_email, password_hash=hashed_password, is_admin=True)
+        db.session.add(new_user)
+        db.session.commit()
+
+
+# append the @auth.admin_required decorator on any route now if you want it to be only accessed by an admin user
+def admin_required(func):
+    @login_required
+    @wraps(func) # necessary, because otherwise decorator would change the name of every function it decorates to "wrapper". See https://stackoverflow.com/questions/308999/what-does-functools-wraps-do
+    def wrapper(*args, **kwargs):
+        if not current_user.is_admin:
+            return login_manager.unauthorized()
+        return func(*args, **kwargs)
+    return wrapper
